requirements.txt
tensorflow
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
keras==3.8.0
numpy==2.0.2
pillow
uvicorn
starlette
scikit-learn